+++ builder_pkgs/mockito/test/deprecated_apis/verify_test.dart
@@ -121,5 +121,16 @@ void main() {
       verify(mock.methodWithPositionalArgs(
           typed(argThat(greaterThanOrEqualTo(100))), 17));
     });
+
+    test('should mock method with mock args', () {
+      var m1 = named(_MockedClass(), name: 'm1');
+      mock.methodWithObjArgs(m1);
+      expectFail(
+          'No matching calls. All calls: _MockedClass.methodWithObjArgs(m1)\n'
+          '$noMatchingCallsFooter', () {
+        verify(mock.methodWithObjArgs(_MockedClass()));
+      });
+      verify(mock.methodWithObjArgs(m1));
+    });
   });
 }